refactor: change the package between domain and collection
