+++ src/main/java/lotto/collections/WinningTicket.java
@@ -1,7 +1,7 @@
 package lotto.collections;
 
 import lotto.domain.LottoNumber;
-import lotto.domain.LottoTicket;
+import lotto.domain.RewardType;
 import lotto.service.WinningService;
 
 public class WinningTicket {
