+++ src/main/java/lotto/domain/RewardType.java
@@ -1,4 +1,4 @@
-package lotto.collections;
+package lotto.domain;
 
 import java.util.Arrays;
 import java.util.List;
