+++ src/test/java/lotto/collections/LottoResultTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.collections;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -10,7 +10,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
 import lotto.collections.LottoResult;
-import lotto.collections.RewardType;
+import lotto.domain.RewardType;
 
 public class LottoResultTest {
