+++ src/test/java/lotto/collections/LottoTicketTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.collections;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -13,6 +13,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
+import lotto.collections.LottoTicket;
+import lotto.domain.LottoNumber;
 import lotto.util.AutoLottoNumberGenerator;
 
 public class LottoTicketTest {
