+++ src/test/java/lotto/collections/WinningNumberTest.java
@@ -1,10 +1,12 @@
-package lotto.domain;
+package lotto.collections;
 
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+
+import lotto.domain.LottoNumber;
 
 public class WinningNumberTest {
