+++ src/test/java/lotto/collections/WinningTicketTest.java
@@ -13,7 +13,7 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
 import lotto.domain.LottoNumber;
-import lotto.domain.LottoTicket;
+import lotto.domain.RewardType;
 
 public class WinningTicketTest {
