+++ src/test/java/lotto/domain/MoneyTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
-import lotto.collections.Money;
+import lotto.domain.Money;
 
 public class MoneyTest {
