+++ src/test/java/lotto/domain/RewardTypeTest.java
@@ -1,4 +1,4 @@
-package lotto.collections;
+package lotto.domain;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -11,6 +11,8 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.MethodSource;
+
+import lotto.domain.RewardType;
 
 public class RewardTypeTest {
